[refactor] Use PascalCase and camelCase and use corresponding absolute routes
