+++ user/src/layouts/MainLayout/FlussAppBar.tsx
@@ -1,5 +1,5 @@
 import { AppBar as Navbar, Button, Divider, Toolbar, Typography } from "@material-ui/core";
-import { createStyles, makeStyles, Theme } from "@material-ui/core/styles";
+import { makeStyles, Theme } from "@material-ui/core/styles";
 import Image from "next/image";
 import router from "next/router";
 import React, { FC } from "react";
@@ -44,48 +44,47 @@ const FlussAppBar: FC = () => {
   );
 };
 
-const useStyles = makeStyles((theme: Theme) =>
-  createStyles({
-    brand: {
-      display: "flex",
-      alignItems: "center",
-      cursor: "pointer",
+const useStyles = makeStyles((theme: Theme) => ({
+  brand: {
+    display: "flex",
+    alignItems: "center",
+    cursor: "pointer",
+    marginRight: theme.spacing(3),
+    padding: 0,
+  },
+  logo: {
+    display: "flex",
+    alignItems: "center",
+    marginRight: theme.spacing(2),
+  },
+  title: {
+    flexGrow: 1,
+    display: "none",
+    fontSize: theme.typography.h5.fontSize,
+    color: theme.palette.primary.main,
+    cursor: "pointer",
+    textTransform: "none",
+    [theme.breakpoints.up("sm")]: {
+      display: "block",
+    },
+  },
+  startButtons: {
+    flexGrow: 1,
+    display: "flex",
+    alignItems: "center",
+  },
+  reportsButton: {
+    borderRadius: 10,
+  },
+  endButtons: {
+    display: "flex",
+    flexDirection: "row",
+    alignItems: "center",
+    "& > *:not(:last-child)": {
       marginRight: theme.spacing(3),
-      padding: 0,
     },
-    logo: {
-      display: "flex",
-      alignItems: "center",
-      marginRight: theme.spacing(2),
-    },
-    title: {
-      flexGrow: 1,
-      display: "none",
-      fontSize: theme.typography.h5.fontSize,
-      color: theme.palette.primary.main,
-      cursor: "pointer",
-      textTransform: "none",
-      [theme.breakpoints.up("sm")]: {
-        display: "block",
-      },
-    },
-    startButtons: {
-      flexGrow: 1,
-      display: "flex",
-      alignItems: "center",
-    },
-    reportsButton: {
-      borderRadius: 10,
-    },
-    endButtons: {
-      display: "flex",
-      flexDirection: "row",
-      alignItems: "center",
-      "& > *:not(:last-child)": {
-        marginRight: theme.spacing(3),
-      },
-    },
-  })
-);
+  },
+}));
 
 export default FlussAppBar;
+
