+++ user/src/layouts/index.tsx
@@ -1,9 +1,10 @@
 import React, { FC } from "react";
 
-import MainLayout from "./main-layout";
+import MainLayout from "./MainLayout";
 
 const Layout: FC = ({ children }) => {
   return <MainLayout>{children}</MainLayout>;
 };
 
 export default Layout;
+
