+++ user/src/pages/_document.tsx
@@ -2,7 +2,7 @@ import { ServerStyleSheets } from "@material-ui/core/styles";
 import Document, { Head, Html, Main, NextScript } from "next/document";
 import React from "react";
 
-import theme from "../styles/theme";
+import theme from "styles/theme";
 
 class FlussDocument extends Document {
   render() {
@@ -64,3 +64,4 @@ FlussDocument.getInitialProps = async (ctx) => {
 };
 
 export default FlussDocument;
+
