+++ user/tests/pages/Home.test.tsx
@@ -2,7 +2,7 @@ import "@testing-library/jest-dom/extend-expect";
 
 import { render } from "@testing-library/react";
 
-import Home from "../../src/pages";
+import Home from "pages";
 
 test("Check for Getting Started Text", () => {
   const { getByText } = render(<Home />);
